requirements.txt
openai
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
requests
python-dateutil
